Finished the functionality of the New Invoice (Main Menu sub window).
It is now able to create new invoices, add items and then save them to the database. (Save and Cancel buttons were added, along with total sale price display window)

+++ Group6FinalProject/Group6FinalProject/clsItem.cs
@@ -10,7 +10,7 @@ namespace Group6FinalProject
     {
         public string ItemCode { get; set; }
         public string ItemDescription { get; set; }
-        public decimal ItemPrice { get; set; }
+        public int ItemPrice { get; set; }
 
         /// <summary>
         /// The to String override will dispaly the item description in the combo Box
